# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Repair Shop Database"
  description: "Mock mechanic shop db"
  version: "1.2.0"
schemes:
- "http"
host: "127.0.0.1:5000"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /service_tickets/my-tickets:
    get:
      tags:
        - Service Tickets
        - Method Protection → Customer Token
      summary: "Get all service tickets for customer - PROTECTED METHOD - add 'Bearer ' and paste in the customer token"
      description: "Customer route to view their service tickets."

      security:
        - bearerAuth: []

      responses:
        200:
          description: "Customer's service tickets shown in a successful response"
          schema:
            $ref: "#/definitions/ViewServiceTickets"
            example:
              application/json:
                [
                  {
                    "id": 3,
                    "inventories":
                      [
                        {
                          "id": 1,
                          "product_name": "Regular Oil - HOUSE",
                          "price": 13.45
                        }
                      ],
                    "mechanics":
                      [
                        {
                          "id": 1,
                          "name": "ed",
                          "phone": "9719719711"
                        },
                        {
                        "id": 2,
                        "name": "edd",
                        "phone": "5035035033"
                        }
                      ],
                    "service_date": "2021-02-20",
                    "service_desc": "oil change",
                    "vin": "abcdefghijklmnop3"
                  },
                  {
                    "id": 2,
                    "vin": "abcdefghijklmnop2",
                    "service_date": "2025-07-17",
                    "service_desc": "oil change",
                    "customer_id": 1,
                    "inventories": [],
                    "mechanics": []
                  }
                ]
        404:
          description: "Service not found"
          schema:
            example:
              {
                "error": "Service tickets not found. It's entirely possible this customer has no tickets."
              }

  # ============== Mechanic Token Protected ==============
  /service_tickets/:
    post:
      tags:
        - Service Tickets
        - Method Protection → Mechanic Token
      summary: "Create a service ticket"
      description: "Required fields: vin (exactly 17 characters long), service_date (format: YYYY-MM-DD), service_desc, customer_id"
      
      security:
        - bearerAuth: []

      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      
      responses:
        201:
          description: "Service ticket creation successful"
          schema:
            $ref: "#/definitions/CreateServiceTicketResponse"
            example:
              application/json:
                {
                  "id": 16,
                  "vin": "abcdefghijklmnop9",
                  "service_date": "2022-11-11",
                  "service_desc": "windshield fluid refill",
                  "customer_id": 1,
                  "inventories": [],
                  "mechanics": []
                }
        400:
          description: "Required data missing from body"
          schema: 
            $ref: "#/definitions/Customer400Message"
            example:
              application/json:
                {
                  "message": "'error details' - all service ticket fields required"
                }
        404:
          description: "Customer not found"
          schema:
            example:
              {
                "error": "Customer not found."
              }
        422:
          description: "Can't make a service ticket for a deleted customer"
          schema:
            example:
              {
                "error": "Customer was previously deleted."
              }
    get:
      tags:
        - Service Tickets
        - Method Protection → Mechanic Token
      summary: "Get All Service Tickets - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanics/Admin only\n *Example* /service_tickets/?page=1&per_page=5"

      security:
        - bearerAuth: []
      
      parameters:
        - in: query
          name: page
          type: integer
          required: false
        - in: query
          name: per_page
          type: integer
          required: false
      
      responses:
        200:
          description: "All service tickets view for mechanics, pagination is optional, "
          schema:
            $ref: "#/definitions/ViewServiceTickets"
            example:
              application/json:
                [
                  {
                    "id": 1,
                    "vin": "abcdefghijklmnop1",
                    "service_date": "2022-11-11",
                    "service_desc": "windshield fluid refill",
                    "customer_id": 1,
                    "inventories": [],
                    "mechanics": []
                  },
                  {
                    "id": 2,
                    "vin": "abcdefghijklmnop2",
                    "service_date": "2022-11-12",
                    "service_desc": "oil change",
                    "customer_id": 2,
                    "inventories": [],
                    "mechanics": []
                  }
                ]
  /service_tickets/{ticket_id}:
    get:
      tags:
        - Service Tickets
        - Method Protection → Mechanic Token
      summary: "Get service ticket by id number - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanic route to get service ticket by id"

      security:
        - bearerAuth: []

      parameters:
        - in: path
          name: ticket_id
          type: integer
          required: true
          description: "The service ticket ID number"

      responses:
        200:
          description: "Service ticket data shown in a successful response"
          schema:
            $ref: "#/definitions/CreateServiceTicketResponse"
            example:
              application/json:
                {
                  "id": 1,
                  "vin": "abcdefghijklmnop1",
                  "service_date": "2022-11-11",
                  "service_desc": "windshield fluid refill",
                  "customer_id": 1,
                  "inventories": [],
                  "mechanics": []
                }
        404:
          description: "Service ticket not found"
          schema:
            example:
              {
                "error": "Service ticket not found."
              }
  /service_tickets/{ticket_id}/edit-mechanics:
    put:
      tags:
        - Service Tickets
        - Method Protection → Mechanic Token
      summary: "Assign mechanics to service ticket by id number - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanic route to assign mechanics to service ticket by id"

      security:
        - bearerAuth: []

      parameters:
        - in: path
          name: ticket_id
          type: integer
          required: true
          description: "The service ticket ID number"
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/EditMechanicsOnServiceTicket"

      responses:
        200:
          description: "Service ticket data shown in a successful response"
          schema:
            $ref: "#/definitions/ViewServiceTickets"
        400:
          description: "Mechanic not found or already assigned to this service ticket"
          schema:
            example:
              {
                "message": "'error details' - both add_mechanic_ids, and remove_mechanic_id fields are required."
              }
...
definitions: 
...
  CreateServiceTicketPayload:
    type: object
    properties:
      vin:
        type: string
        minLength: 17
        maxLength: 17
      service_date:
        type: string
        format: date-time
      service_desc:
        type: string
      customer_id:
        type: integer
    example:
      vin: "abcdefghijklmnop9"
      service_date: "2022-11-11"
      service_desc: "windshield fluid refill"
      customer_id: 1
...
  CreateServiceTicketResponse:
    type: object
    properties:
      id:
        type: integer
      vin:
        type: string
        minLength: 17
        maxLength: 17
      service_date:
        type: string
        format: date
      service_desc:
        type: string
      customer_id:
        type: integer
      inventories:
        type: array
        items:
          $ref: "#/definitions/ViewInventoryAsCustomer"
      mechanics:
        type: array
    example:
      {
        "id": 16,
        "customer_id": 1,
        "inventories": [],
        "mechanics": [],
        "service_date": "2022-11-11",
        "service_desc": "windshield fluid refill",
        "vin": "abcdefghijklmnop9"
      }
...
  EditMechanicsOnServiceTicket:
    type: object
    properties:
      add_mechanic_ids:
        type: array
        items:
          type: integer
      remove_mechanic_id:
        type: integer
    example:
      add_mechanic_ids: [1, 2]
      remove_mechanic_id: [3]
...
  ViewInventoryAsCustomer:
    type: object
    description: "Inventory item view for customers"
    properties:
      product_name:
        type: string
      price:
        type: number
        format: float
    example:
      {
        "product_name": "Regular Oil - HOUSE",
        "price": 13.45
      }
...
  ViewServiceTickets:
    type: object
    properties:
      id:
        type: integer
      vin:
        type: string
        minLength: 17
        maxLength: 17
      service_date:
        type: string
        format: date
      service_desc:
        type: string
      inventories:
        type: array
        items:
          $ref: "#/definitions/InventoriesOnServiceTicket"
      mechanics:
        type: array
        items:
          $ref: "#/definitions/MechanicsOnServiceTicket"
    example:
      {
        "id": 16,
        "inventories":
          [
            {
              "product_name": "Regular Oil - HOUSE",
              "price": 13.45
            },
            {
              "product_name": "High Mileage Oil - HOUSE",
              "price": 15.95
            }
          ],
        "mechanics":
          [
            {
              "name": "ed",
              "phone": "9719719711",
              "email": "ed@repairshop.com"
            },
            {
              "name": "edd",
              "phone": "5035035033",
              "email": "edd@repairshop.com"
            }
          ],
        "service_date": "2022-11-11",
        "service_desc": "windshield fluid refill",
        "vin": "abcdefghijklmnop9"
      }
  InventoriesOnServiceTicket:
    type: array
    items:
      $ref: "#/definitions/ViewInventoryAsCustomer"
    example:
      [
        {
          "product_name": "Regular Oil - HOUSE",
          "price": 13.45
        },
        {
          "product_name": "High Mileage Oil - HOUSE",
          "price": 15.95
        }
      ]
  MechanicsOnServiceTicket:
    type: array
    items:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
    example:
      [
        {
          "name": "ed",
          "phone": "9719719711",
          "email": "ed@repairshop.com"
        },
        {
          "name": "edd",
          "phone": "5035035033",
          "email": "edd@repairshop.com"
        }
      ]
  Customer400Message:
    type: object
    properties: 
      message:
        type: string
    example:
      message: "'error details' - all customer data fields required."
...
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header